move extract_H_ranges to new module
