+++ libensemble/tests/unit_tests/test_sim_dir_properties.py
@@ -3,20 +3,21 @@ import shutil
 import numpy as np
 from libensemble.output_directory import EnsembleDirectory
 from libensemble.utils.loc_stack import LocationStack
+from libensemble.utils.misc import extract_H_ranges
 
 
 def test_range_single_element():
     """Single H_row labeling"""
 
     work = {"H_fields": ["x", "num_nodes", "procs_per_node"], "libE_info": {"H_rows": np.array([5]), "workerID": 1}}
-    assert EnsembleDirectory.extract_H_ranges(work) == "5", "Failed to correctly parse single H row"
+    assert extract_H_ranges(work) == "5", "Failed to correctly parse single H row"
 
 
 def test_range_two_separate_elements():
     """Multiple H_rows, non-sequential"""
 
     work = {"H_fields": ["x", "num_nodes", "procs_per_node"], "libE_info": {"H_rows": np.array([2, 8]), "workerID": 1}}
-    assert EnsembleDirectory.extract_H_ranges(work) == "2_8", "Failed to correctly parse nonsequential H rows"
+    assert extract_H_ranges(work) == "2_8", "Failed to correctly parse nonsequential H rows"
 
 
 def test_range_two_ranges():
@@ -26,7 +27,7 @@ def test_range_two_ranges():
         "H_fields": ["x", "num_nodes", "procs_per_node"],
         "libE_info": {"H_rows": np.array([0, 1, 2, 3, 7, 8]), "workerID": 1},
     }
-    assert EnsembleDirectory.extract_H_ranges(work) == "0-3_7-8", "Failed to correctly parse multiple H ranges"
+    assert extract_H_ranges(work) == "0-3_7-8", "Failed to correctly parse multiple H ranges"
 
 
 def test_range_mixes():
@@ -37,7 +38,7 @@ def test_range_mixes():
         "libE_info": {"H_rows": np.array([2, 3, 4, 6, 8, 9, 11, 14]), "workerID": 1},
     }
     assert (
-        EnsembleDirectory.extract_H_ranges(work) == "2-4_6_8-9_11_14"
+        extract_H_ranges(work) == "2-4_6_8-9_11_14"
     ), "Failed to correctly parse H row single elements and ranges."
