+++ libensemble/utils/misc.py
@@ -1,0 +1,23 @@
+"""
+Misc internal functions
+"""
+
+from itertools import groupby
+from operator import itemgetter
+
+
+def extract_H_ranges(Work):
+    """Convert received H_rows into ranges for labeling"""
+    work_H_rows = Work["libE_info"]["H_rows"]
+    if len(work_H_rows) == 1:
+        return str(work_H_rows[0])
+    else:
+        # From https://stackoverflow.com/a/30336492
+        ranges = []
+        for diff, group in groupby(enumerate(work_H_rows.tolist()), lambda x: x[0] - x[1]):
+            group = list(map(itemgetter(1), group))
+            if len(group) > 1:
+                ranges.append(str(group[0]) + "-" + str(group[-1]))
+            else:
+                ranges.append(str(group[0]))
+        return "_".join(ranges)
